Replace lift-json with json4s-jackson.
The aim of the Json4s project is to provide a common API for
Scala JSON libraries.  It is Apache-licensed, easier for
downstream distributions to package, and mostly API-compatible
with lift-json.  Furthermore, the Jackson-backed implementation
parses faster than lift-json on all but the smallest inputs.

Backported patch from master to 0.9.0.

Conflicts:
	core/src/main/scala/org/apache/spark/deploy/master/ui/ApplicationPage.scala
	core/src/main/scala/org/apache/spark/deploy/master/ui/IndexPage.scala
	core/src/main/scala/org/apache/spark/ui/JettyUtils.scala
	core/src/test/scala/org/apache/spark/deploy/JsonProtocolSuite.scala