Ensure assertions in Graph.apply are asserted.
The Graph.apply test in GraphSuite had some assertions in a closure in
a graph transformation. This caused two problems:

1. because assert() was called, test classes were reachable from the
closures, which made them not serializable, and

2. (more importantly) these assertions never actually executed, since
they occurred within a lazy map()

This commit simply changes the Graph.apply test to collects the graph
triplets so it can assert about each triplet from a map method.